+++ test/platform/swift/unit_integration/mocks/MockLoggerBridging.swift
@@ -64,6 +64,10 @@ extension MockLoggerBridging: LoggerBridging {
 
     public func getDeviceID() -> String { "deviceID" }
 
+    public func getSdkStatus() -> SdkStatus {
+        SdkStatus(initializationState: .notStarted, lastHandshakeTime: nil, lastConfigDeliveryTime: nil)
+    }
+
     public func log(
         level: LogLevel,
         message: @autoclosure () -> String,